+++ index.js
@@ -45,6 +45,7 @@ var CleverTap = {
     CleverTapInboxDidInitialize: CleverTapReact.CleverTapInboxDidInitialize,
     CleverTapInboxMessagesDidUpdate: CleverTapReact.CleverTapInboxMessagesDidUpdate,
     CleverTapInboxMessageButtonTapped: CleverTapReact.CleverTapInboxMessageButtonTapped,
+    CleverTapInboxMessageTapped:CleverTapReact.CleverTapInboxMessageTapped,
     CleverTapDisplayUnitsLoaded: CleverTapReact.CleverTapDisplayUnitsLoaded,
     CleverTapInAppNotificationButtonTapped: CleverTapReact.CleverTapInAppNotificationButtonTapped,
     CleverTapFeatureFlagsDidUpdate: CleverTapReact.CleverTapFeatureFlagsDidUpdate,
@@ -106,10 +107,16 @@ var CleverTap = {
     /**
     * Manually set the push token on the CleverTap user profile
     * @param {string} token - the device token
-    * @param {string} type - for Android only, specifying the type of push service token. Values can be CleverTap.FCM for Firebase or CleverTap.XPS for Xiaomi or CleverTap.BPS for Baidu or CleverTap.HPS for Huawei
+    * @param {string} type - for Android only, specifying the type of push service token. Values can be CleverTap.FCM for Firebase or CleverTap.XPS for Xiaomi or CleverTap.BPS for Baidu or CleverTap.HPS for Huawei,
+     * @param {string} region - for xps only ,...
     */
-    setPushToken: function (token, type) {
-        CleverTapReact.setPushTokenAsString(token, type);
+    setPushToken: function (token, type,region="") { //"abc", "fcm"
+        if(type === CleverTap.XPS){
+            CleverTapReact.setPushTokenAsString(token, type,region);
+        }
+        else {
+            CleverTapReact.setPushTokenAsString(token, type);
+        }
     },
 
     /**